#134 [refact] 카프카 관련 패키지 구조를 변경한다.
