+++ src/main/java/com/dataracy/modules/behaviorlog/adapter/kafka/provider/BehaviorLogKafkaProducerAdapter.java
@@ -1,4 +1,4 @@
-package com.dataracy.modules.behaviorlog.adapter.message.kafka;
+package com.dataracy.modules.behaviorlog.adapter.kafka.provider;
 
 import com.dataracy.modules.behaviorlog.application.port.out.BehaviorLogSendProducerPort;
 import com.dataracy.modules.behaviorlog.domain.model.BehaviorLog;
